+++ src/embeds/components/Items/index.js
@@ -28,6 +28,7 @@ type Props = {
   className?: string,
   mode?: 'rune' | 'item',
   statIds: { [key: number]: number },
+  optionalText?: string,
 };
 
 @connect(mapStateToProps, {
@@ -37,16 +38,22 @@ type Props = {
 export default class ItemsEmbed extends Component {
   props: Props;
 
-  static renderItem (id: number, mode?: 'rune' | 'item', statId?: number, items?: Items, itemStats?: ItemStats) {
-    const item = items && items[id];
+  static renderItem (
+    id: number,
+    mode?: 'rune' | 'item',
+    statId?: number,
+    items?: Items,
+    itemStats?: ItemStats,
+    optionalText?: string,
+  ) {
     const selectedStat = statId && itemStats && itemStats[statId];
-
-    // no item to render, skip.
+    const item = items && items[id];
     if (!item) {
       return undefined;
     }
 
-    // Apply stat on item.
+    // TODO: Move this into a custom reducer.
+    // See: https://github.com/madou/armory-react/issues/243
     if (selectedStat && item.details && !item.details.infix_upgrade_applied) {
       const attributes = applyAttributes(item, selectedStat);
 
@@ -58,13 +65,19 @@ export default class ItemsEmbed extends Component {
       item.details.infix_upgrade_applied = true;
     }
 
-    return (<Item
-      key={id}
-      item={item}
-      name={mode === 'rune' ? 'Rune' : undefined}
-      tooltipType={mode === 'rune' ? 'amulets' : undefined}
-      className={styles.item}
-    />);
+    if (id > -1) {
+      return (
+        <Item
+          key={id}
+          item={item}
+          name={mode === 'rune' ? 'Rune' : undefined}
+          tooltipType={mode === 'rune' ? 'amulets' : undefined}
+          className={styles.item}
+        />
+      );
+    }
+
+    return <Item tooltipTextOverride={optionalText} />;
   }
 
   componentWillMount () {
@@ -75,11 +88,18 @@ export default class ItemsEmbed extends Component {
   }
 
   render () {
-    const { ids, statIds, items, itemStats, className, mode } = this.props;
+    const { ids, statIds, items, itemStats, className, mode, optionalText } = this.props;
 
     return (
       <div className={className}>
-        {ids.map((id) => ItemsEmbed.renderItem(id, mode, statIds[id], items, itemStats))}
+        {ids.map((id) => ItemsEmbed.renderItem(
+          id,
+          mode,
+          statIds[id],
+          items,
+          itemStats,
+          optionalText,
+        ))}
       </div>
     );
   }