+++ src/embeds/components/Items/index.spec.js
@@ -1,0 +1,25 @@
+describe('<Items /> embed', () => {
+  context('when mounting', () => {
+    it('should fetch items', () => {
+
+    });
+
+    it('should fetch itemStats', () => {
+
+    });
+  });
+
+  describe('rendering', () => {
+    it('should render nothing for items that have no data', () => {
+
+    });
+
+    it('should render an empty item for negative ids', () => {
+
+    });
+
+    it('should render items that have data', () => {
+
+    });
+  });
+});