+++ src/features/Character/components/Skill/index.js
@@ -12,13 +12,14 @@ import styles from './styles.less';
 type Props = {
   data?: {},
   className?: string,
+  tooltipTextOverride?: string,
 };
 
-const Skill = ({ data, className }: Props) => {
+const Skill = ({ data, className, tooltipTextOverride }: Props) => {
   const error = get(data, 'error');
 
   return (
-    <TooltipTrigger type="skill" data={data || 'No Skill Selected'}>
+    <TooltipTrigger type="skill" data={tooltipTextOverride || data || 'No Skill Selected'}>
       {(error || !data)
         ? <EmptySkill />
         : (